+++ lib/ws_servers/api/factories/create_dms_control.js
@@ -1,0 +1,13 @@
+const DmsRemoteControl = require('../dms_remote_control')
+
+/**
+ * @param {APIWSServer} server
+ * @returns {DmsRemoteControl}
+ */
+module.exports = (server) => {
+  const { hostedURL, restURL } = server
+  return new DmsRemoteControl({
+    hostedURL,
+    restURL
+  })
+}
