+++ lib/ws_servers/api/start_connections.js
@@ -2,13 +2,12 @@
 
 const { DEFAULT_USER } = require('../../constants')
 
-const openAuthBitfinexConnection = require('./open_auth_bitfinex_connection')
+const createClient = require('./open_auth_bitfinex_connection')
+const createAlgoWorker = require('./factories/create_algo_worker')
+const createDmsControl = require('./factories/create_dms_control')
+const createStrategyManager = require('./factories/create_strategy_manager')
 const getUserSettings = require('../../util/user_settings')
 const capture = require('../../capture')
-const AlgoWorker = require('./algos/algo_worker')
-const send = require('../../util/ws/send')
-const DmsRemoteControl = require('./dms_remote_control')
-const StrategyManager = require('./handlers/strategy/strategy_manager')
 
 class ConnectionManager {
   constructor () {
@@ -79,8 +78,7 @@ class ConnectionManager {
       return
     }
 
-    const { hostedURL, restURL } = server
-    dmsControl = new DmsRemoteControl({ hostedURL, restURL })
+    dmsControl = createDmsControl(server)
     ws.setDmsControl(dmsControl)
 
     const { apiKey, apiSecret } = this._getCredentials()
@@ -127,7 +125,7 @@ class ConnectionManager {
       return
     }
 
-    algoWorker = await this._createAlgoWorker(server, ws)
+    algoWorker = await createAlgoWorker(server, ws)
     ws.setAlgoWorker(algoWorker)
 
     await algoWorker.start({ apiKey, apiSecret, userId })
@@ -136,21 +134,16 @@ class ConnectionManager {
   /**
    * @private
    */
-  _startBfxClient ({ ws, d, apiKey, apiSecret, wsURL, restURL, mode, isPaper }, shouldUpdate = false) {
+  _startBfxClient ({ ws, ...args }, shouldUpdate = false) {
     let client = ws.getClient()
     if (client && client.isOpen && !shouldUpdate) {
       return
     }
 
-    client = openAuthBitfinexConnection({
+    client = createClient({
+      ...args,
       ws,
-      d,
-      dms: false,
-      apiKey,
-      apiSecret,
-      wsURL,
-      restURL,
-      isPaper
+      dms: false
     })
 
     ws.setClient(client)
@@ -167,56 +160,8 @@ class ConnectionManager {
       return
     }
 
-    const { wsURL, restURL, strategyExecutionDB } = server
-    const bcast = { ws: send.bind(send, ws) }
-
-    strategyManager = new StrategyManager(
-      {
-        dms: false,
-        wsURL,
-        restURL
-      },
-      bcast,
-      strategyExecutionDB
-    )
+    strategyManager = createStrategyManager(server, ws)
     ws.setStrategyManager(strategyManager)
-  }
-
-  /**
-   * @param {APIWSServer} server
-   * @param {Session} ws
-   * @returns {Promise<AlgoWorker>}
-   */
-  async _createAlgoWorker (server, ws) {
-    const {
-      tracerDir,
-      wsURL,
-      restURL,
-      algos,
-      algoDB,
-      logAlgoOpts,
-      marketData
-    } = server
-    const bcast = { ws: send.bind(send, ws) }
-    const { affiliateCode } = await server.getUserSettings()
-
-    return new AlgoWorker(
-      {
-        dms: false,
-        affiliateCode,
-        wsURL,
-        restURL,
-        signalTracerOpts: {
-          enabled: true,
-          dir: tracerDir
-        }
-      },
-      algos,
-      bcast,
-      algoDB,
-      logAlgoOpts,
-      marketData
-    )
   }
 }
